+++ egui_glow/src/painter.rs
@@ -35,7 +35,8 @@ pub struct Painter {
     is_embedded: bool,
     vertex_array: crate::misc_util::VAO,
     srgb_support: bool,
-    scale_filter: u32,
+    /// The filter used for subsequent textures.
+    texture_filter: u32,
     post_process: Option<PostProcess>,
     vertex_buffer: glow::Buffer,
     element_array_buffer: glow::Buffer,
@@ -54,22 +55,22 @@ pub struct Painter {
 }
 
 #[derive(Copy, Clone)]
-pub enum ScaleFilter {
+pub enum TextureFilter {
     Linear,
     Nearest,
 }
 
-impl Default for ScaleFilter {
+impl Default for TextureFilter {
     fn default() -> Self {
-        ScaleFilter::Linear
+        TextureFilter::Linear
     }
 }
 
-impl ScaleFilter {
+impl TextureFilter {
     fn glow_code(&self) -> u32 {
         match self {
-            ScaleFilter::Linear => glow::LINEAR,
-            ScaleFilter::Nearest => glow::NEAREST,
+            TextureFilter::Linear => glow::LINEAR,
+            TextureFilter::Nearest => glow::NEAREST,
         }
     }
 }
@@ -212,7 +213,7 @@ impl Painter {
                 is_embedded: matches!(shader_version, ShaderVersion::Es100 | ShaderVersion::Es300),
                 vertex_array,
                 srgb_support,
-                scale_filter: ScaleFilter::default().glow_code(),
+                texture_filter: TextureFilter::default().glow_code(),
                 post_process,
                 vertex_buffer,
                 element_array_buffer,
@@ -247,7 +248,7 @@ impl Painter {
                 gl,
                 self.is_webgl_1,
                 self.srgb_support,
-                self.scale_filter,
+                self.texture_filter,
                 &pixels,
                 font_image.width,
                 font_image.height,
@@ -413,8 +414,8 @@ impl Painter {
 
     // Set the filter to be used for any subsequent textures loaded via
     // `set_texture`.
-    pub fn set_scale_filter(&mut self, scale_filter: ScaleFilter) {
-        self.scale_filter = scale_filter.glow_code();
+    pub fn set_texture_filter(&mut self, texture_filter: TextureFilter) {
+        self.texture_filter = texture_filter.glow_code();
     }
 
     // ------------------------------------------------------------------------
@@ -440,7 +441,7 @@ impl Painter {
             gl,
             self.is_webgl_1,
             self.srgb_support,
-            self.scale_filter,
+            self.texture_filter,
             &pixels,
             image.size[0],
             image.size[1],